move relations endpoint to plant layer
