+++ backend/src/controller/mod.rs
@@ -3,6 +3,7 @@
 pub mod config;
 pub mod layers;
 pub mod map;
+pub mod plant_layer;
 pub mod planting_suggestions;
 pub mod plantings;
 pub mod plants;
