+++ backend/src/model/entity.rs
@@ -2,6 +2,7 @@
 
 pub mod layer_impl;
 pub mod map_impl;
+pub mod plant_layer;
 pub mod plantings;
 pub mod plantings_impl;
 pub mod plants_impl;
