+++ backend/src/service/plants.rs
@@ -8,8 +8,6 @@ use crate::config::data::AppDataInner;
 use crate::model::dto::Page;
 use crate::model::dto::PageParameters;
 use crate::model::dto::PlantSuggestionsSearchParameters;
-use crate::model::dto::RelationSearchParameters;
-use crate::model::dto::RelationsDto;
 use crate::{
     error::ServiceError,
     model::{
@@ -36,19 +34,6 @@ pub async fn find(
         _ => Plants::find_any(page_parameters, &mut conn).await?,
     };
 
-    Ok(result)
-}
-
-/// Get all relations of a certain plant.
-///
-/// # Errors
-/// If the connection to the database could not be established.
-pub async fn find_relations(
-    search_query: RelationSearchParameters,
-    app_data: &Data<AppDataInner>,
-) -> Result<RelationsDto, ServiceError> {
-    let mut conn = app_data.pool.get().await?;
-    let result = Plants::find_relations(search_query, &mut conn).await?;
     Ok(result)
 }
